+++ java/corecomponents/daf.etl/src/main/java/net/atos/daf/ct2/etl/common/bo/Trip.java
@@ -20,9 +20,9 @@ public class Trip implements Serializable {
 	private String vin;
 	private Long startDateTime;
 	private Long endDateTime;
-	private Integer gpsTripDist;
+	private Long gpsTripDist;
 	private Long tripCalDist;
-	private Integer vIdleDuration;
+	private Long vIdleDuration;
 	private Double vGrossWeightCombination;
 	private Double tripCalAvgSpeed;
 	private Long gpsStartVehDist;
@@ -31,9 +31,9 @@ public class Trip implements Serializable {
 	private Double gpsStartLongitude;
 	private Double gpsEndLatitude;
 	private Double gpsEndLongitude;
-	private Integer vUsedFuel;
+	private Long vUsedFuel;
 	private Long tripCalUsedFuel;
-	private Integer vTripMotionDuration;
+	private Long vTripMotionDuration;
 	private Long tripCalDrivingTm;
 	private Long receivedTimestamp;
 	private Double tripCalC02Emission;
@@ -48,9 +48,9 @@ public class Trip implements Serializable {
 	private Double tripCalCrsCntrlDistAbv75;
 	private Double tripCalAvgTrafficClsfn;
 	private Double tripCalCCFuelConsumption;
-	private Integer vCruiseControlFuelConsumed;
-	private Integer vCruiseControlDist;
-	private Integer vIdleFuelConsumed;
+	private Long vCruiseControlFuelConsumed;
+	private Long vCruiseControlDist;
+	private Long vIdleFuelConsumed;
 	private Double tripCalfuelNonActiveCnsmpt;
 	private Double tripCalDpaScore;
 	private String driverId;
